+++ tests/e2e/mod.rs
@@ -1,3 +1,4 @@
 pub mod caldav;
 pub mod carddav;
 pub mod util;
+pub mod webdav;